LoRa: Skip RX2 if RX1 has not received timeout
With low datarates radio might be already receiving on RX1 when RX2 should be
opened. In this case RX2 should not be opened but RX1 should be kept open
until all data has been received.

This commit changes behaviour so that RX2 is opened only if RX1 has either
timed out or data has been received.